Split out gm_1832
