+++ src/melee/gm/gm_1832.c
@@ -1,0 +1,634 @@
+#include "gm_17C0.h"
+
+#include "gm_1601.h"
+
+#include <dolphin/gx.h>
+#include <math_ppc.h>
+#include <sysdolphin/baselib/gobjproc.h>
+#include <sysdolphin/baselib/tobj.h>
+#include <sysdolphin/baselib/util.h>
+#include <melee/cm/camera.h>
+#include <melee/ef/efasync.h>
+#include <melee/ef/eflib.h>
+#include <melee/ft/ft_0877.h>
+#include <melee/ft/ftbosslib.h>
+#include <melee/ft/ftlib.h>
+#include <melee/gm/gm_1A3F.h>
+#include <melee/gm/gm_1A45.h>
+#include <melee/gm/types.h>
+#include <melee/gr/ground.h>
+#include <melee/gr/grpushon.h>
+#include <melee/gr/stage.h>
+#include <melee/it/item.h>
+#include <melee/lb/lb_00B0.h>
+#include <melee/lb/lb_00F9.h>
+#include <melee/lb/lbaudio_ax.h>
+#include <melee/lb/lbbgflash.h>
+#include <melee/mp/mpcoll.h>
+#include <melee/pl/player.h>
+#include <melee/sc/types.h>
+
+static HSD_JObj* lbl_804735A8[0x10];
+static struct {
+    HSD_ImageDesc x40[3];
+    HSD_ImageDesc x88[3];
+    u8 pad[0xE0 - 0x88 - 3*0x18];
+    u8 xE0;
+    u8 xE1;
+    u8 xE2;
+    u8 xE3;
+    int xE4;
+    int xE8;
+    u8 xEC;
+    u8 xED;
+    u8 xEE;
+    u8 xEF;
+    u8 xF0;
+    u8 xF1;
+} lbl_804735E8;
+
+static struct {
+    u8 pad[0x6A8];
+    Vec3 x630[4];
+    Vec3 x650[4];
+    Vec3 x6A8[1];
+}* lbl_804D6604;
+
+/// #fn_8018325C
+
+/// #fn_80184138
+
+/// #fn_801849E0
+
+/// #fn_80184A04
+
+/// #fn_80184A28
+
+/// #fn_80184A4C
+
+/// #fn_80184A70
+
+/// #fn_80184A94
+
+/// #fn_80184AB8
+
+/// #fn_8018504C
+
+/// #fn_801851C0
+
+void fn_801852FC(HSD_GObj* gobj)
+{
+    int i;
+    int j;
+
+    PAD_STACK(0x18);
+
+    if (HSD_CObjSetCurrent(GET_COBJ(gobj))) {
+        for (i = 0; i < lbl_804735E8.xE0; i++) {
+            for (j = 0; j < lbl_804735E8.xE0 + 1; j++) {
+                if (j == 0 || i + 1 != j) {
+                    Player_800366DC(j, 0);
+                } else {
+                    Player_800366DC(j, 1);
+                }
+            }
+            Camera_800313E0(gobj, 1);
+            HSD_StateInvalidate(-1);
+            HSD_ImageDescCopyFromEFB(&lbl_804735E8.x40[i], 0x82, 0, 0, 0);
+            HSD_ImageDescCopyFromEFB(&lbl_804735E8.x88[i], 0x82, 0, 1, 1);
+        }
+        HSD_CObjEndCurrent();
+    }
+    lbl_804735E8.xE1 = 1;
+}
+
+void fn_80185408(int x, float arg8, float arg9, float argA, float argB)
+{
+    u8 _[0x30];
+    Mtx sp1C;
+    GXSetNumChans(1);
+    GXSetChanCtrl(GX_COLOR0A0, 0, GX_SRC_REG, GX_SRC_VTX, 0, GX_DF_NONE, GX_AF_NONE);
+    GXSetNumTexGens(0);
+    GXSetZMode(1, GX_LEQUAL, 1);
+    GXSetNumTevStages(1);
+    GXSetTevOrder(GX_TEVSTAGE0, GX_TEXCOORD_NULL, GX_TEXMAP_NULL, GX_COLOR_NULL);
+    GXSetTevOp(GX_TEVSTAGE0, GX_PASSCLR);
+    GXSetCullMode(GX_CULL_NONE);
+    GXSetAlphaCompare(GX_ALWAYS, 0, GX_AOP_OR, GX_ALWAYS, 0);
+    GXSetBlendMode(GX_BM_NONE, GX_BL_SRCALPHA, GX_BL_INVSRCALPHA, GX_LO_CLEAR);
+    GXSetColorUpdate(0);
+    GXClearVtxDesc();
+    GXSetVtxAttrFmt(GX_VTXFMT0, GX_VA_POS, GX_TEX_ST, GX_RGBA6, 0);
+    GXSetVtxDesc(GX_VA_POS, GX_DIRECT);
+    GXLoadPosMtxImm(HSD_identityMtx, 0);
+    GXSetCurrentMtx(0);
+    MTXOrtho(sp1C, 0.0F, 480.0F, 0.0F, 640.0F, 0.0F, 5000.0F);
+    GXSetProjection(sp1C, GX_ORTHOGRAPHIC);
+    GXBegin(GX_QUADS, GX_VTXFMT0, 4);
+    GXWGFifo.f32 = argA;
+    GXWGFifo.f32 = arg8;
+    GXWGFifo.f32 = -4932.0F;
+    GXWGFifo.f32 = argB;
+    GXWGFifo.f32 = arg8;
+    GXWGFifo.f32 = -4932.0F;
+    GXWGFifo.f32 = argB;
+    GXWGFifo.f32 = arg9;
+    GXWGFifo.f32 = -4932.0F;
+    GXWGFifo.f32 = argA;
+    GXWGFifo.f32 = arg9;
+    GXWGFifo.f32 = -4932.0F;
+    GXSetColorUpdate(1);
+    HSD_StateInvalidate(-1);
+}
+
+extern float MSL_TrigF_80400770[];
+extern float MSL_TrigF_80400774[];
+double fn_801855BC(double arg8)
+{
+    f64 temp_f2;
+    f64 temp_f2_2;
+    f64 temp_f2_3;
+    f64 temp_f2_4;
+
+    if (arg8 > 0.0) {
+        temp_f2_2 = __frsqrte(arg8);
+        temp_f2_3 = 0.5 * temp_f2_2 * -(((f64) arg8 * (temp_f2_2 * temp_f2_2)) - 3.0);
+        temp_f2_4 = 0.5 * temp_f2_3 * -(((f64) arg8 * (temp_f2_3 * temp_f2_3)) - 3.0);
+        temp_f2 = 0.5 * temp_f2_4 * -(((f64) arg8 * (temp_f2_4 * temp_f2_4)) - 3.0);
+        return arg8 * (0.5 * temp_f2 * -(((f64) arg8 * (temp_f2 * temp_f2)) - 3.0));
+    }
+    if (0.0 == arg8) {
+        return 0.0;
+    }
+    if (arg8) {
+        return MSL_TrigF_80400770[0];
+    }
+    return MSL_TrigF_80400774[0];
+}
+
+void fn_8018564C(HSD_GObj* gobj)
+{
+    fn_80185408(0xFC8000, 0.0F, 480.0F, 200.0F, 640.0F);
+    HSD_SObjLib_803A54EC(gobj, 7);
+}
+
+void fn_8018569C(HSD_GObj* gobj)
+{
+    int i;
+    PAD_STACK(8);
+    if (HSD_CObjSetCurrent(GET_COBJ(gobj))) {
+        for (i = 0; i < lbl_804735E8.xE0 + 1; i++) {
+            if (i == 0) {
+                Player_800366DC(i, 1);
+            } else {
+                Player_800366DC(i, 0);
+            }
+        }
+        Camera_800313E0(gobj, 1);
+        gobj->gxlink_prios = 0x1001;
+        HSD_GObj_80390ED0(gobj, 7);
+        HSD_CObjEndCurrent();
+    }
+}
+
+void fn_8018575C(HSD_GObj* gobj)
+{
+    if (HSD_CObjSetCurrent(GET_COBJ(gobj))) {
+        Camera_800313E0(gobj, 1);
+        gobj->gxlink_prios = 0x1001;
+        HSD_GObj_80390ED0(gobj, 7);
+        HSD_CObjEndCurrent();
+    }
+}
+
+/// #fn_801857C4
+
+/// #fn_801859C8
+
+/// #fn_80185A0C
+
+/// #fn_80185D64
+
+/// #fn_80185E34
+
+/// #fn_80185F5C
+
+/// #fn_80186080
+
+/// #fn_801861B8
+
+static SceneDesc* lbl_804D65FC;
+
+void fn_80186400(void)
+{
+    HSD_GObj* gobj = GObj_Create(0xE, 0xF, 0);
+    HSD_JObj* jobj = HSD_JObjLoadJoint(lbl_804D65FC->models[0]->joint);
+    HSD_GObjObject_80390A70(gobj, HSD_GObj_804D7849, jobj);
+    GObj_SetupGXLink(gobj, HSD_GObj_JObjCallback, 0xB, 0);
+    HSD_GObjProc_8038FD54(gobj, fn_80184AB8, 0);
+    gm_8016895C(jobj, lbl_804D65FC->models[0], 0);
+    HSD_JObjReqAnimAll(jobj, 0.0F);
+    HSD_JObjAnimAll(jobj);
+    if (lbl_804735E8.xE4 != 3) {
+        lb_80011E24(jobj, &lbl_804735A8[1], 4, -1);
+        lb_80011E24(jobj, &lbl_804735A8[2], 5, -1);
+        lb_80011E24(jobj, &lbl_804735A8[3], 0x3A, -1);
+        lb_80011E24(jobj, &lbl_804735A8[4], 0x3B, -1);
+        HSD_JObjSetTranslateZ(lbl_804735A8[3], 10000.0F);
+        HSD_JObjSetTranslateZ(lbl_804735A8[4], 10000.0F);
+    }
+}
+
+/// #fn_80186634
+
+/// #gm_80186DFC_OnFrame
+
+/// #gm_80186E30_OnEnter
+
+/// #fn_80186EFC
+
+/// #fn_80186F6C
+
+/// #fn_801873F0
+
+/// #fn_80187494
+
+/// #fn_801874FC
+
+void fn_80187714(void)
+{
+    Camera_80028B9C(6);
+    lb_8000FCDC();
+    mpColl_80041C78();
+    Ground_801C0378(64);
+    Stage_802251E8(0, NULL);
+    Stage_8022524C();
+    Item_80266FA8();
+    Item_80266FCC();
+    efLib_8005B4B8();
+    efAsync_8006737C(0);
+}
+
+/// #gm_8018776C_OnFrame
+
+/// #gm_801877A8_OnEnter
+
+/// #fn_80187910
+
+/// #fn_80187AB4
+
+/// #fn_80187C9C
+
+/// #fn_80187CF4
+
+/// #gm_80187F48_OnEnter
+
+/// #gm_80188364_OnLeave
+
+/// #gm_8018838C_OnFrame
+
+/// #gm_801883C0
+
+/// #gm_8018841C
+
+/* 4D4190 */ static s32 lbl_804D4190 = 0xFFFFFFFF;
+
+struct lbl_803D9D20_t {
+    /*  +0 */ char pad_0[0xB0];
+    /* +B0 */ HSD_CObj* cobj;
+    /* +B4 */ char pad_B4[0xD8 - 0xB4];
+    /* +D8 */ Vec3 eye_position;
+    /* +E4 */ char pad_E4[0xEC - 0xE4];
+    /* +EC */ Vec3 interest;
+};
+
+static struct lbl_803D9D20_t lbl_803D9D20 = { 0 };
+
+static int lbl_80473700[69];
+static TmData gm_804771C4;
+
+int gm_80188454(int idx)
+{
+    return lbl_80473700[idx + 2];
+}
+
+/// #fn_8018846C
+
+/// #fn_801884F8
+
+/// #fn_80188550
+
+/// #fn_80188644
+
+/// #fn_80188738
+
+/// #fn_80188910
+
+/// #fn_80188B3C
+
+/// #fn_80188D3C
+
+/// #fn_80188EE8
+
+/// #fn_801891F4
+
+/// #fn_80189B88
+
+/// #gm_80189CDC
+
+/// #fn_8018A000
+
+/// #gm_8018A160
+
+/// #gm_8018A188
+
+/// #gm_8018A1D8
+
+/// #gm_8018A228
+
+/// #gm_8018A25C
+
+/// #gm_8018A290
+
+/// #gm_8018A2C4
+
+/// #gm_8018A314
+
+/// #fn_8018A364
+
+/// #fn_8018A514
+
+/// #fn_8018A970
+
+/// #fn_8018AA74
+
+/// #fn_8018B090
+
+/// #fn_8018C8D4
+
+/// #fn_8018D50C
+
+/// #fn_8018DC18
+
+/// #fn_8018DF68
+
+/// #fn_8018E46C
+
+/// #fn_8018E618
+
+/// #fn_8018E85C
+
+/// #fn_8018EC48
+
+/// #fn_8018EC7C
+
+/// #fn_8018ECA8
+
+/// #fn_8018F00C
+
+/// #gm_8018F1B0
+
+/// #fn_8018F310
+
+u8 fn_8018F3BC(s32 arg0)
+{
+    return M2C_FIELD(&lbl_803D9D20.pad_0[arg0], u8*, 0x59);
+}
+
+/// #fn_8018F3D0
+
+/// #fn_8018F410
+
+/// #fn_8018F4A0
+
+/// #fn_8018F508
+
+/// #fn_8018F5F0
+
+s32 fn_8018F62C(HSD_GObj* arg0)
+{
+    return M2C_FIELD(arg0, s32*, 0x2C);
+}
+
+TmData* gm_8018F634(void)
+{
+    return &gm_804771C4;
+}
+
+/// #fn_8018F640
+
+/// #fn_8018F674
+
+/// #fn_8018F6A8
+
+/// #fn_8018F6DC
+
+/// #fn_8018F6FC
+
+/// #fn_8018F71C
+
+/// #fn_8018F74C
+
+/// #fn_8018F808
+
+/// #fn_8018F888
+
+/// #fn_8018FA24
+
+void fn_8018FBD8(void* arg0, s32 arg1)
+{
+    M2C_FIELD(arg0, s32*, 0x2C) = arg1;
+}
+
+/// #fn_8018FBE0
+
+/// #fn_8018FDC4
+
+/// #fn_8018FF9C
+
+/// #fn_80190174
+
+/// #fn_801901F8
+
+/// #fn_8019027C
+
+/// #fn_801902F0
+
+/// #fn_8019035C
+
+/// #fn_8019044C
+
+/// #fn_80190480
+
+void fn_801904D0(void)
+{
+    HSD_CObjSetInterest(lbl_803D9D20.cobj, &lbl_803D9D20.interest);
+    HSD_CObjSetEyePosition(lbl_803D9D20.cobj, &lbl_803D9D20.eye_position);
+}
+
+/// #fn_80190520
+
+/// #gm_801905F0
+
+/// #fn_80190ABC
+
+/// #gm_80190EA4
+
+/// #gm_80190FE4
+
+/// #fn_801910E0
+
+/// #fn_80191154
+
+/// #fn_80191240
+
+/// #fn_801913BC
+
+/// #fn_80191678
+
+/// #fn_8019175C
+
+/// #fn_801918F0
+
+/// #fn_80191A54
+
+/// #fn_80191B5C
+
+/// #fn_80191CA4
+
+/// #fn_80191D38
+
+/// #fn_80191E9C
+
+/// #fn_80191FD4
+
+/// #fn_8019237C
+
+/// #fn_8019249C
+
+/// #fn_80192690
+
+/// #fn_80192758
+
+/// #fn_80192938
+
+/// #fn_80192BB0
+
+/// #fn_80192E6C
+
+/// #fn_80193230
+
+/// #fn_80193308
+
+/// #fn_801935B8
+
+/// #fn_801937C4
+
+/// #fn_80193B58
+
+/// #fn_80193FCC
+
+/// #fn_80194658
+
+/// #fn_801949B4
+
+/// #fn_80194BC4
+
+/// #fn_80194D84
+
+/// #fn_80194F30
+
+/// #fn_801953C8
+
+/// #fn_80195AF0
+
+/// #fn_80195CCC
+
+/// #fn_8019610C
+
+/// #gm_8019628C_OnFrame
+
+/// #gm_801963B4_OnEnter
+
+/// #gm_801964A4_OnLeave
+
+/// #fn_80196510
+
+s32 fn_8019655C(void)
+{
+    return lbl_804D4190;
+}
+
+/// #fn_80196564
+
+/// #fn_80196594
+
+/// #fn_801965C4
+
+/// #fn_80196684
+
+/// #fn_801967E0
+
+/// #fn_80196CF8
+
+/// #fn_80196DBC
+
+/// #fn_80196E30
+
+/// #fn_80196EEC
+
+/// #fn_80196FFC
+
+/// #fn_801973F8
+
+/// #fn_801975C8
+
+/// #fn_801976D4
+
+/// #fn_801977AC
+
+/// #fn_80197AF0
+
+/// #fn_80197D4C
+
+/// #fn_80197E18
+
+/// #fn_80197FD8
+
+/// #fn_801981A0
+
+/// #fn_801983E4
+
+/// #fn_80198584
+
+/// #fn_801985D4
+
+/// #fn_80198824
+
+/// #fn_80198BA0
+
+/// #fn_80198C60
+
+/// #fn_80198D18
+
+/// #fn_80198EBC
+
+/// #fn_80199AF0
+
+/// #fn_8019A158
+
+/// #fn_8019A71C
+
+/// #gm_8019A828
+
+/// #fn_8019A86C
+
+/// #fn_8019AF50
+
+/// #gm_8019B2DC_OnFrame
+
+/// #fn_8019B458
+
+/// #fn_8019B81C
+
+/// #fn_8019B860
